changes made

+++ src/Simulation/Sink.java
@@ -59,7 +59,7 @@ public class Sink implements ProductAcceptor
 		int[] tmp = new int[numbers.size()];
 		for (int i=0; i < numbers.size(); i++)
 		{
-			tmp[i] = (numbers.get(i)).intValue();
+			tmp[i] = numbers.get(i);
 		}
 		return tmp;
 	}
@@ -70,7 +70,7 @@ public class Sink implements ProductAcceptor
 		double[] tmp = new double[times.size()];
 		for (int i=0; i < times.size(); i++)
 		{
-			tmp[i] = (times.get(i)).doubleValue();
+			tmp[i] = times.get(i);
 		}
 		return tmp;
 	}
